one-pixel-attack/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-image>=0.21.0
tqdm>=4.65.0

# Medical imaging specific
torchxrayvision>=1.2.0

# Additional utilities
matplotlib>=3.7.0
ipython>=8.0.0

# Database and environment
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0